refactor(ios): shrink LMWF parser/import/export services (#347)
Behavior-preserving refactors behind the parser's spec-backed test suites;
no suppressions:

- MarkdownParser (781 lines) split along construct seams into extensions:
  +Structure (workout header/section), +Exercises (exercise blocks, groups),
  +Sets (set lines, modifiers merge), +Modifiers (rpe/rest/tempo/flags);
  long bodies decomposed (parseWorkout, parseExerciseBlock,
  parseGroupedExercises c12->3, parseSetLine c11->2). Error/warning emission
  order and all message strings preserved verbatim.
- MarkdownParserHelpers: shared parseRepsOrTimeValue extracted (was
  duplicated verbatim across three set parsers); parseModifiersAndTrailingText
  c13->9 via parseFlagModifier.
- JsonImportService: long import bodies decomposed; 7-param helper grouped
  into a small context struct. AnthropicService request builder split.
  WorkoutExportService / SessionLMWFEncoder / WorkoutGenerationService /
  WorkoutHistoryService: complexity extracted into per-concern helpers.
  Export/encoder output byte-identical (round-trip suites pass).

Baseline regenerated: 68 -> 50 entries, removals only. Parser family:
143 tests pass; full unit suite: 914 tests, 0 failures.

Part of #351. Closes #347.
